# api/openapi/authserver.yaml
openapi: 3.1.0
info:
  title: Keystone Auth Server API
  description: |
    OAuth 2.0 / OpenID Connect Identity Provider for Keystone IAM.
    Implements RFC 6749, RFC 7636 (PKCE S256), RFC 9068 (JWT profile),
    RFC 9700 (OAuth 2.0 Security BCP), and append-only tamper-evident audit logging.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8080
    description: Local development Auth Server
  - url: http://localhost:9100
    description: Local development Admin/Health port

paths:
  /.well-known/openid-configuration:
    get:
      summary: OpenID Connect Discovery Document
      description: Returns OIDC provider configuration metadata according to RFC 8414 and OIDC Discovery 1.0.
      operationId: getOpenIDConfiguration
      responses:
        '200':
          description: OIDC Discovery Configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenIDConfiguration'

  /jwks.json:
    get:
      summary: JSON Web Key Set (JWKS)
      description: Returns active public keys used to verify RS256 JWT access and ID tokens (RFC 7517).
      operationId: getJWKS
      responses:
        '200':
          description: JWKS public keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWKS'

  /authorize:
    get:
      summary: Authorization Endpoint
      description: |
        Interactive OAuth 2.0 authorization endpoint initiating Authorization Code Flow with mandatory PKCE (S256).
        Redirects to /login if unauthenticated, to /consent if consent required, or redirects back to redirect_uri with code.
      operationId: authorize
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          required: false
          schema:
            type: string
            default: "openid profile email"
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
        - name: prompt
          in: query
          required: false
          schema:
            type: string
            enum: [none, login, consent]
      responses:
        '302':
          description: Redirect to login, consent, or back to client redirect_uri with authorization code or error.
        '400':
          description: Invalid client or invalid redirect URI (renders error page without redirecting).

  /token:
    post:
      summary: Token Endpoint
      description: |
        Issues access tokens (RFC 9068 at+jwt), ID tokens, and rotating refresh tokens.
        Supported grant types: authorization_code (with PKCE), refresh_token (with rotation and reuse detection), and client_credentials.
      operationId: issueToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Successful token response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: OAuth 2.0 RFC 6749 error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '401':
          description: Invalid client authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'

  /userinfo:
    get:
      summary: UserInfo Endpoint
      description: Returns OpenID Connect standard claims for the authenticated subject.
      operationId: getUserInfo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Standard OIDC claims
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: Invalid, expired, or missing Bearer token

  /login:
    get:
      summary: Interactive Login Page
      description: Renders HTML username and password login form with CSRF token.
      responses:
        '200':
          description: HTML login form
          content:
            text/html:
              schema:
                type: string
    post:
      summary: Submit Login Credentials
      description: Verifies credentials against Argon2id hash and establishes session cookie.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password, csrf_token]
              properties:
                username:
                  type: string
                password:
                  type: string
                csrf_token:
                  type: string
                return_to:
                  type: string
      responses:
        '302':
          description: Redirects to return_to or home on success.
        '400':
          description: Invalid form or CSRF token.
        '401':
          description: Invalid credentials.

  /register:
    get:
      summary: Interactive Registration Page
      description: Renders HTML user registration form with CSRF token.
      responses:
        '200':
          description: HTML registration form
          content:
            text/html:
              schema:
                type: string
    post:
      summary: Submit Registration
      description: Creates user account with Argon2id password hash.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, email, password, csrf_token]
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                csrf_token:
                  type: string
                return_to:
                  type: string
      responses:
        '302':
          description: Redirects to login or return_to on success.
        '400':
          description: Validation error or duplicate username/email.

  /consent:
    get:
      summary: Consent Screen
      description: Displays requested OAuth scopes and prompts user for approval.
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: scope
          in: query
          required: true
          schema:
            type: string
        - name: return_to
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HTML consent form
          content:
            text/html:
              schema:
                type: string
    post:
      summary: Submit Consent Decision
      description: Records user scope consent and redirects back to authorization flow.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [client_id, scope, decision, csrf_token, return_to]
              properties:
                client_id:
                  type: string
                scope:
                  type: string
                decision:
                  type: string
                  enum: [allow, deny]
                csrf_token:
                  type: string
                return_to:
                  type: string
      responses:
        '302':
          description: Redirects back to return_to.

  /healthz:
    get:
      summary: Liveness Probe
      description: Basic HTTP check returning 200 OK if service is alive.
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"

  /readyz:
    get:
      summary: Readiness Probe
      description: Verifies active database connection, connection pool, and migrations.
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ready"
                  database:
                    type: string
                    example: "ok"
        '503':
          description: Service is not ready (e.g., database unreachable)

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    OpenIDConfiguration:
      type: object
      required:
        - issuer
        - authorization_endpoint
        - token_endpoint
        - jwks_uri
        - userinfo_endpoint
        - response_types_supported
        - subject_types_supported
        - id_token_signing_alg_values_supported
        - code_challenge_methods_supported
      properties:
        issuer:
          type: string
          example: "http://localhost:8080"
        authorization_endpoint:
          type: string
          example: "http://localhost:8080/authorize"
        token_endpoint:
          type: string
          example: "http://localhost:8080/token"
        jwks_uri:
          type: string
          example: "http://localhost:8080/jwks.json"
        userinfo_endpoint:
          type: string
          example: "http://localhost:8080/userinfo"
        response_types_supported:
          type: array
          items:
            type: string
          example: ["code"]
        subject_types_supported:
          type: array
          items:
            type: string
          example: ["public"]
        id_token_signing_alg_values_supported:
          type: array
          items:
            type: string
          example: ["RS256"]
        code_challenge_methods_supported:
          type: array
          items:
            type: string
          example: ["S256"]
        grant_types_supported:
          type: array
          items:
            type: string
          example: ["authorization_code", "client_credentials", "refresh_token"]
        token_endpoint_auth_methods_supported:
          type: array
          items:
            type: string
          example: ["client_secret_basic", "client_secret_post", "none"]

    JWKS:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            $ref: '#/components/schemas/JWK'

    JWK:
      type: object
      required: [kty, use, alg, kid, n, e]
      properties:
        kty:
          type: string
          example: "RSA"
        use:
          type: string
          example: "sig"
        alg:
          type: string
          example: "RS256"
        kid:
          type: string
          example: "key_01jh12345678"
        n:
          type: string
          description: Base64url-encoded modulus
        e:
          type: string
          description: Base64url-encoded exponent
          example: "AQAB"

    TokenRequest:
      type: object
      required: [grant_type]
      properties:
        grant_type:
          type: string
          enum: [authorization_code, refresh_token, client_credentials]
        code:
          type: string
        redirect_uri:
          type: string
        client_id:
          type: string
        code_verifier:
          type: string
        refresh_token:
          type: string
        scope:
          type: string

    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: RS256 signed JWT with RFC 9068 typ=at+jwt
        token_type:
          type: string
          example: "Bearer"
        expires_in:
          type: integer
          example: 3600
        refresh_token:
          type: string
          description: High-entropy single-use refresh token (rotated on each use)
        id_token:
          type: string
          description: RS256 signed OIDC ID token
        scope:
          type: string
          example: "openid profile email"

    OAuthError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "invalid_grant"
        error_description:
          type: string
          example: "Authorization code has expired or was already redeemed"

    UserInfoResponse:
      type: object
      required: [sub]
      properties:
        sub:
          type: string
          format: uuid
        preferred_username:
          type: string
          example: "alice"
        email:
          type: string
          format: email
          example: "alice@example.com"
        email_verified:
          type: boolean
          example: false